feat: Add image erosion + dilation of sky-extracted images
The purpose of erosion followed by dilation is to remove any small patches of sky among the treeline which may be extracted when extracting sky from the frames.
If these small patches remain in the picture, they may be mistaken as drones and hence cause false positives.
Hence, the sky-extracted frame is eroded to remove these patches, and then dilated to keep the original image features.
This erosion + dilation process is defined as "morphological opening" in OpenCV