feat: Dynamically determine whether sun compensation is needed
Let the code decide whether to use sun compensation or not, instead of having the user manually specify whether to use it.
This is done by measuring brightness of the frame. The frame is converted to HSV, and the histogram of V used to calculate average brightness.
If average brightness is above a certain threshold, sun compensation will be applied to the frame