feat: Add basic sun compensation into ROS single detector stack
This ports the sun compensation algorithm from the old python code into the ROS stack.

The main changes are:

(1) Modify the average_brightness function to accept different colortype conversions
(2) Add the extract_sky function
(3) Add ROS parameters that are needed for sun compensation

The next step is to add this to the multi detection stack as well.