feat: Further segment sky into blue and non-blue regions
This is inspired from Shao Xuan's method of setting all blue pixels to white
which effectively maxes out the contrast. The sky is further separated into
blue and non-blue regions, and the localised contrast increase is applied to
the blue regions. Two methods (convertTo and directly setting to white) are
available. To identify blue regions, the RGB channels are used.