feat: Combine sky thresholding with saturation method
This ensures that data among the treeline is retained as it is not affected
by the saturation. In addition, the step of setting dark areas to black was
removed as it was tested to be not helpful in the object detection, and
created more false positives among the treeline instead.