feat: Use Otsu Thresholding to segment sky and treeline
Otsu thresholding helps to dynamically determine the threshold V value
so that we don't have to manually set the value on our own. It assumes
a bimodal image - e.g. the gradient change between sky and treeline must
be clear enough. Tested on several videos (including those with dark and
cloudy conditions) and verified that sky and treeline were accurately split