feat: Improvements to contrast and saturation scaling
Contrast scaling threshold is a hardcoded value instead of average brightness;
this is for situations where the entire sky is either too bright or too dark,
which means average brightness would not be effective threshold.

Saturation scaling is based on inverse linear relationship with pixel brightness.

- Max saturation scale factor = 1 (i.e. no saturation change)
- Min saturation scale factor = 0.3
- Min pixel brightness = 1 (brightness of 0 is ignored as those are masked-out areas)
- Max pixel brightness = 255

Thus, at max pixel brightness, a saturation scale factor of 0.3 is used.
This linear model is designed by ourselves to approsimate how much saturation to lower
and is not in existing literature