fix: Major bug fixes in the trackplot code
Too many changes that I can remember, but the most obvious one
was pertaining to the cross correlation code (which is different
from cross coefficient). A lot of typo errors were also fixed, and
errors that were carried over from Python's over simplification (and
a large part owing to the loss implicit typecasts).