feat: Port single cam env compensation code onto reid_develop work
This ports the environmental compensation work (formerly called sun compensation)
from the sunlight-union branch onto the top of the reid_develop work. This is to
avoid the messy job of merging branches and resolving the conflicts, as significant
changes were made to the code for both branches. This commit only ports work done
on the single camera code.